# amplify/backend/function/PermissionFunction/src/swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Identity and Permission Management API",
    "description": "Handles staff permissions and roles",
    "version": "1.0.0"
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/v1/permissions/staff/{identifier}": {
      "get": {
        "description": "Get all or check the presense of a set of given permissions",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The staff identifier provided by cognito"
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/v1/permissions/roles/staff/{identifier}": {
      "get": {
        "description": "Get all or check the presense of a set of given roles",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The staff identifier provided by cognito"
          },
          {
            "name": "roles",
            "in": "body",
            "description": "An array of role names to test the specified user against",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {}
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/v1/permissions/{identifier}": {
      "get": {
        "description": "Get a specific permission from a given name",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The permission to retreive"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/v1/permissions/role/{identifier}": {
      "get": {
        "description": "Get a specific role and its permissions from a given name",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The role to retreive"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/v1/permissions/{identifier}/members": {
      "get": {
        "description": "Get staff associated with a specific permission. [Supports Pagnetation]",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The permission to retreive members from"
          },
          {
            "name": "size",
            "in": "query",
            "description": "Size of returned page",
            "required": false,
            "type": "string"
          },
          {
            "name": "index",
            "in": "query",
            "description": "Index of returned page",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "416": {
            "description": "Range Not Satisfiable"
          }
        }
      }
    },
    "/v1/permissions/role/{identifier}/members": {
      "get": {
        "description": "Get staff associated with a specific role [Supports Pagnetation]",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The role to retreive members from"
          },
          {
            "name": "size",
            "in": "query",
            "description": "Size of returned page",
            "required": false,
            "type": "string"
          },
          {
            "name": "index",
            "in": "query",
            "description": "Index of returned page",
            "required": false,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "416": {
            "description": "Range Not Satisfiable"
          }
        }
      }
    },
    "/v1/permissions/staff/{identifier}/grant": {
      "put": {
        "description": "Grant a specific set of permissions to a given user",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The staff member grant permissions to"
          },
          {
            "name": "permissions",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "permissions": {
                  "example": "any"
                }
              }
            },
            "description": "An array of permission strings",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/v1/permissions/staff/{identifier}/revoke": {
      "put": {
        "description": "Revoke a specific set of permissions from given user",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The staff member revoke permissions from"
          },
          {
            "name": "permissions",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "permissions": {
                  "example": "any"
                }
              }
            },
            "description": "An array of permission strings",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/v1/permissions/roles/staff/{identifier}/grant": {
      "put": {
        "description": "Grant a specific set of roles to a given user",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The staff member grant roles to"
          },
          {
            "name": "roles",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "roles": {
                  "example": "any"
                }
              }
            },
            "description": "An array of role strings",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/v1/permissions/roles/staff/{identifier}/revoke": {
      "put": {
        "description": "Revoke a specific set of roles from given user",
        "parameters": [
          {
            "name": "identifier",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "roles": {
                  "example": "any"
                }
              }
            },
            "required": true,
            "description": "An array of role strings"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/v1/permissions/roles/{identifier}": {
      "put": {
        "description": "Create or Update a given role",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The role to update"
          },
          {
            "name": "description",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "description": {
                  "example": "any"
                }
              }
            },
            "description": "The updated description string",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "409": {
            "description": "Conflict"
          }
        }
      },
      "delete": {
        "description": "Delete a given role",
        "parameters": [
          {
            "name": "identifier",
            "in": "path",
            "required": true,
            "type": "string",
            "description": "The role to delete"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  }
}